fix(mongodb): Incremental data INCENTIVE_APP_CP_INVNT_PURCHS_LINE is missing invntIndex.catalogItem
Update events, merge arrays, and build arrayFilters from before first. If there is a js node in the middle to change the field name, the data in before will not be renamed. As a result, arrayFilters cannot correctly obtain the field value, and instead takes the value from after first.

Closes TAP-3414